pyproject.toml
[project]
name = "babs"
version = "0.1.0"
description = "Polymarket algorithmic trading bot using the RBI methodology"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
  "py-clob-client>=0.0.1",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "ccxt>=4.0.0",
  "ta>=0.10.0",
  "click>=8.1.0",
]

[project.scripts]
babs = "babs.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/babs"]
